requirements.txt
# Core dependencies
openai~=1.75.0
python-dotenv~=1.1.0

# Data processing
pandas==2.2.3
numpy~=2.2.4
openpyxl==3.1.5

# Machine Learning and NLP
sentence-transformers~=4.0.2
scikit-learn~=1.6.1
transformers~=4.51.2

# Evaluation and visualization
matplotlib~=3.8.0
seaborn~=0.13.0

# Web scraping (for existing checker functionality)
selenium~=4.24.0

# Optional: LangChain components (if needed for advanced RAG)
langchain-community~=0.3.21
langchain-text-splitters~=0.3.8
langchain-core~=0.3.54
langchain~=0.3.23
langchain-openai~=0.3.14

# Text processing
jieba~=0.42.1
rank-bm25~=0.2.2

# Development and testing
pytest~=7.4.0
pytest-cov~=4.1.0
black~=23.0.0
flake8~=6.0.0

# Logging and monitoring
loguru~=0.7.0
tqdm~=4.66.0